Added FolderCleanserService to FolderCleanserConsole. Service retrieves paths from API, loops through retrieving all files from each and processessing. (doesn't currently delete). Posts summary history for each path back to API. changed type from int to double for 2 SummaryHistoryModel properties (FileSizeDeletedMB and ProcessingDurationMins) as well as associated locations (back end model, SQL table).

+++ FolderCleanserSolution/FolderCleanserBackEndLibrary/Models/SummaryHistoryModel.cs
@@ -6,7 +6,7 @@ public class SummaryHistoryModel
     public int PathId { get; set; }
     public DateTime ProcessingStartDateTime { get; set; }
     public DateTime ProcessingEndDateTime { get; set; }
-    public int ProcessingDurationMins { get; set; }
+    public double ProcessingDurationMins { get; set; }
     public int FilesDeletedCount { get; set; }
     public double FileSizeDeletedMB { get; set; }
 }
