+++ FolderCleanserSolution/FolderCleanserDB/dbo/Tables/SummaryHistory.sql
@@ -4,7 +4,7 @@
     [PathId] INT NOT NULL, 
     [ProcessingStartDateTime] DATETIME2 NOT NULL, 
     [ProcessingEndDateTime] DATETIME2 NOT NULL, 
-    [ProcessingDurationMins] INT NOT NULL DEFAULT 0, 
+    [ProcessingDurationMins] FLOAT NOT NULL DEFAULT 0, 
     [FilesDeletedCount] INT NOT NULL DEFAULT 0, 
     [FileSizeDeletedMB] FLOAT NOT NULL DEFAULT 0
 )
